requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.4
django-environ==0.12.0
google-auth==2.43.0
google-genai==1.49.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mysqlclient==2.2.7
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sweetify==2.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.11.0
